@@ -23,15 +23,10 @@
       "homepage": "http://avalanche123.com"
     }
   ],
-  "require": {
-    "php": ">=5.6.0"
-  },
   "require-dev": {
-    "behat/behat": "~3.0.6",
-    "phpunit/php-code-coverage": "~2.0",
-    "phpunit/php-token-stream": "~1.3",
-    "phpunit/phpunit": "~4.8",
-    "symfony/process": "~2.1"
+    "phpunit/phpunit": "^9",
+    "behat/behat": "^3.8",
+    "symfony/process": "^5.2"
   },
   "config": {
     "bin-dir": "bin/"
