@@ -21,7 +21,7 @@ namespace Cassandra;
 /**
  * @requires extension cassandra
  */
-class DateTest extends \PHPUnit_Framework_TestCase
+class DateTest extends \PHPUnit\Framework\TestCase
 {
     const SECONDS_PER_DAY = 86400;
 
@@ -40,7 +40,8 @@ class DateTest extends \PHPUnit_Framework_TestCase
     public function testConstructNow()
     {
         $date = new Date();
-        $this->assertEquals($date->seconds(), (int) (time() / self::SECONDS_PER_DAY) * self::SECONDS_PER_DAY, "", 1);
+        $this->assertEquals($date->seconds(), (int) (time() / self::SECONDS_PER_DAY) * self::SECONDS_PER_DAY, "");
+        $this->assertEqualsWithDelta($date->seconds(), (int) (time() / self::SECONDS_PER_DAY) * self::SECONDS_PER_DAY, 1, "");
     }
 
     public function testFromDateTime()
