@@ -21,7 +21,7 @@ namespace Cassandra;
 /**
  * @requires extension cassandra
  */
-class TimeTest extends \PHPUnit_Framework_TestCase
+class TimeTest extends \PHPUnit\Framework\TestCase
 {
     public function testConstruct()
     {
@@ -38,24 +38,21 @@ class TimeTest extends \PHPUnit_Framework_TestCase
     public function testConstructNow()
     {
         $time = new Time();
-        $this->assertEquals($time->seconds(), time() % (24 * 60 * 60), "", 1);
+        $this->assertEquals($time->seconds(), time() % (24 * 60 * 60), "");
+        $this->assertEqualsWithDelta($time->seconds(), time() % (24 * 60 * 60), 1, "");
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage nanoseconds must be nanoseconds since midnight, -1 given
-     */
     public function testConstructNegative()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('nanoseconds must be nanoseconds since midnight, -1 given');
         $time = new Time(-1);
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage nanoseconds must be nanoseconds since midnight, '86400000000000' given
-     */
     public function testConstructTooBig()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('nanoseconds must be nanoseconds since midnight, \'86400000000000\' given');
         $time = new Time("86400000000000");
     }
 
