@@ -23,7 +23,7 @@ use Cassandra\Type;
 /**
  * @requires extension cassandra
  */
-class MapTest extends \PHPUnit_Framework_TestCase
+class MapTest extends \PHPUnit\Framework\TestCase
 {
     public function testDefinesMapType()
     {
@@ -53,36 +53,25 @@ class MapTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(0, count($map));
     }
 
-    /**
-     * @expectedException        InvalidArgumentException
-     * @expectedExceptionMessage Not enough values, maps can only be created
-     *                           from an even number of values, where each odd
-     *                           value is a key and each even value is a value,
-     *                           e.g create(key, value, key, value, key, value)
-     */
     public function testPreventsCreatingMapWithoutEnoughValues()
     {
+        $this->expectException(\InvalidArgumentException::class);
         Type::map(Type::varchar(), Type::int())
             ->create("a", 1, "b", 2, "c", 3, "d", 4, "e");
     }
 
-    /**
-     * @expectedException        InvalidArgumentException
-     * @expectedExceptionMessage argument must be a string, 1 given
-     */
     public function testPreventsCreatingMapWithUnsupportedTypes()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('argument must be a string, 1 given');
         Type::map(Type::varchar(), Type::int())
             ->create(1, "a");
     }
 
-    /**
-     * @expectedException        InvalidArgumentException
-     * @expectedExceptionMessage keyType must be a valid Cassandra\Type,
-     *                           an instance of Cassandra\Type\UnsupportedType given
-     */
     public function testPreventsDefiningMapsWithUnsupportedTypes()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('keyType must be a valid Cassandra\Type, an instance of Cassandra\Type\UnsupportedType given');
         Type::map(new UnsupportedType(), Type::varchar());
     }
 
