@@ -23,7 +23,7 @@ use Cassandra\Type;
 /**
  * @requires extension cassandra
  */
-class TupleTest extends \PHPUnit_Framework_TestCase
+class TupleTest extends \PHPUnit\Framework\TestCase
 {
     public function testDefinesTupleType()
     {
@@ -62,25 +62,20 @@ class TupleTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($tuple->get(2), null);
     }
 
-    /**
-     * @expectedException        InvalidArgumentException
-     * @expectedExceptionMessage argument must be a string, 1 given
-     */
     public function testPreventsCreatingTupleWithInvalidType()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('argument must be a string, 1 given');
         Type::tuple(Type::varchar())->create(1);
     }
 
-    /**
-     * @expectedException        InvalidArgumentException
-     * @expectedExceptionMessage type must be a valid Cassandra\Type,
-     *                           an instance of Cassandra\Type\UnsupportedType given
-     */
     public function testPreventsDefiningTuplesWithUnsupportedTypes()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('type must be a valid Cassandra\Type, an instance of Cassandra\Type\UnsupportedType given');
         Type::tuple(new UnsupportedType());
     }
-
+    
     /**
      * @dataProvider equalTypes
      */
