@@ -21,14 +21,12 @@ namespace Cassandra;
 /**
  * @requires extension cassandra
  */
-class UserTypeValueTest extends \PHPUnit_Framework_TestCase
+class UserTypeValueTest extends \PHPUnit\Framework\TestCase
 {
-    /**
-     * @expectedException         InvalidArgumentException
-     * @expectedExceptionMessage  Unsupported type 'invalid type'
-     */
     public function testSupportsOnlyCassandraTypes()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('Unsupported type \'invalid type\'');
         new UserTypeValue(array('name1' => 'invalid type'));
     }
 
@@ -37,6 +35,7 @@ class UserTypeValueTest extends \PHPUnit_Framework_TestCase
      */
     public function testSupportsAllCassandraTypes($type)
     {
+        self::expectNotToPerformAssertions();
         new UserTypeValue(array('name1' => $type));
     }
 
@@ -145,33 +144,27 @@ class UserTypeValueTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($udt, $other);
     }
 
-    /**
-     * @expectedException         InvalidArgumentException
-     * @expectedExceptionMessage  Invalid name 'invalid'
-     */
     public function testSetInvalidName()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('Invalid name \'invalid\'');
         $udt = new UserTypeValue(array('name1' => Type::int()));
         $udt->set('invalid', 42);
     }
 
-    /**
-     * @expectedException         InvalidArgumentException
-     * @expectedExceptionMessage  Invalid name 'invalid'
-     */
     public function testGetInvalidName()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('Invalid name \'invalid\'');
         $udt = new UserTypeValue(array('name1' => Type::int()));
         $udt->set('name1', 42);
         $udt->get('invalid');
     }
 
-    /**
-     * @expectedException         InvalidArgumentException
-     * @expectedExceptionMessage  argument must be an int, 'text' given
-     */
     public function testInvalidType()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('argument must be an int, \'text\' given');
         $udt = new UserTypeValue(array('name1' => Type::int()));
         $udt->set('name1', 'text');
     }
