@@ -18,6 +18,8 @@
 
 namespace Cassandra;
 
+use Cassandra\Exception\UnavailableException;
+
 /**
  * Retry policy integration tests.
  */
@@ -54,7 +56,7 @@ class RetryPolicyIntegrationTest extends BasicIntegrationTest {
     /**
      * Setup the retry policy for multiple nodes.
      */
-    public function setUp() {
+    public function setUp(): void {
         // Ensure RF = 3 (anything greater than 1 is good)
         $this->replicationFactor = 3;
 
@@ -224,11 +226,11 @@ class RetryPolicyIntegrationTest extends BasicIntegrationTest {
      * @ticket PHP-60
      *
      * @cassandra-version-2.0
-     *
-     * @expectedException \Cassandra\Exception\UnavailableException
-     * @expectedExceptionMessageRegExp |Cannot achieve consistency level .*|
      */
     public function testFallThroughPolicyWrite() {
+        $this->expectException(UnavailableException::class);
+        $this->expectExceptionMessageMatches('|Cannot achieve consistency level .*|');
+
         // Create the retry policy (RF = 3 with 1 node)
         $policy = new RetryPolicy\Fallthrough();
 
@@ -260,11 +262,11 @@ class RetryPolicyIntegrationTest extends BasicIntegrationTest {
      * @ticket PHP-60
      *
      * @cassandra-version-2.0
-     *
-     * @expectedException \Cassandra\Exception\UnavailableException
-     * @expectedExceptionMessageRegExp |Cannot achieve consistency level .*|
      */
     public function testFallThroughPolicyRead() {
+        $this->expectException(UnavailableException::class);
+        $this->expectExceptionMessageMatches('|Cannot achieve consistency level .*|');
+
         // Create the retry policy (RF = 3 with 1 node)
         $policy = new RetryPolicy\Fallthrough();
 
