@@ -18,6 +18,8 @@
 
 namespace Cassandra;
 
+use Cassandra\Exception\InvalidQueryException;
+
 /**
  * Tuple integration tests.
  *
@@ -212,9 +214,10 @@ class TupleIntegrationTest extends CollectionsIntegrationTest {
      *
      * @test
      * @ticket PHP-58
-     * @expectedException \Cassandra\Exception\InvalidQueryException
      */
     public function testInvalidType() {
+        $this->expectException(InvalidQueryException::class);
+
         $validType = Type::tuple(Type::int());
         $invalidType = Type::tuple(Type::varchar());
 
